# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Kanban application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kanban API Documentation
  version: "1.0"
paths:
...
  /organizations/{orgID}/projects:
    get:
      description: Retrieves all projects within a specified organization. User must
        own the organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of projects
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all projects in an organization
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Creates a new project within a specified organization. User must
        own the organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project creation details
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden (if user does not own organization)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found (if organization not found)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict (if project name already exists in organization)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new project
      tags:
      - Projects
  /organizations/{orgID}/projects/{projectID}:
    delete:
      description: Deletes a specific project by its ID within a specified organization.
        User must own the organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a project
      tags:
      - Projects
    get:
      description: Retrieves a specific project by its ID within a specified organization.
        User must own the organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project details
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get project by ID
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates a specific project by its ID within a specified organization.
        User must own the organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Project update details
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a project
      tags:
      - Projects
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your JWT token in the format "Bearer <your_token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...